--- .idea/misc.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ExternalStorageConfigurationManager" enabled="true" />
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_21" project-jdk-name="Kotlin SDK" project-jdk-type="KotlinSDK">
-    <output url="file://$PROJECT_DIR$/out" />
-  </component>
-</project>
